+++ frontend/package.json
@@ -35,8 +35,8 @@
     "@emotion/react": "^11.8.2",
     "@emotion/serialize": "^1.0.2",
     "@emotion/styled": "^11.8.1",
-    "@glideapps/glide-data-grid": "4.1.0",
-    "@glideapps/glide-data-grid-source": "4.1.0",
+    "@glideapps/glide-data-grid": "5.0.0",
+    "@glideapps/glide-data-grid-source": "5.0.0",
     "@loaders.gl/core": "^2.2.8",
     "@loaders.gl/csv": "^2.2.8",
     "@loaders.gl/gltf": "^2.2.8",
@@ -100,7 +100,6 @@
     "remark-math": "^4.0.0",
     "sass": "^1.43.4",
     "sprintf-js": "^1.1.2",
-    "styled-components": "^5.3.3",
     "styletron-engine-atomic": "^1.4.6",
     "styletron-react": "^5.2.6",
     "typed-signals": "^2.2.0",
@@ -145,7 +144,6 @@
     "@types/react-virtualized": "^9.21.11",
     "@types/react-window": "^1.8.3",
     "@types/sprintf-js": "^1.1.2",
-    "@types/styled-components": "^5.1.24",
     "@types/styletron-engine-atomic": "^1.1.0",
     "@types/styletron-react": "^5.0.2",
     "@types/styletron-standard": "^2.0.1",
